Break router into package
Add Response and create packages to support

Add Index route

Move Storage.go and CheckAuthorized

Add VerifyUser handler

Add RefreshCustomToken route

Move helpers to single file

Add InitiateMeetover route

Add SendPush route

Add matches route

Add test route

Move ml directory to root of project

Finish restructuring

Attempt to get ml working

Return MatchResponse

Fix rebase from dev-fixes

Add gorgonia override

Split firebase package out into more files

Add LinkedIn note

Remove unauthorized response

rebase again

Rename ml to backend/data

Wrap up changes after rebase

Missed more code

RequestScreen patch
