+++ backend/firebase/Auth.go
@@ -1,0 +1,21 @@
+package firebase
+
+import (
+	"context"
+	"errors"
+)
+
+// CreateCustomToken Creates firebase based IM access token for the user with LinkedIn user ID
+func CreateCustomToken(ID string) (string, error) {
+	client, err := fbApp.Auth(context.Background())
+	if err != nil {
+		return "", errors.New("error getting Auth client")
+	}
+
+	token, err := client.CustomToken(ID)
+	if err != nil {
+		return "", errors.New("error minting custom token")
+	}
+
+	return token, nil
+}
