+++ backend/location/Location.go
@@ -1,4 +1,4 @@
-package main
+package location
 
 import (
 	"fmt"
