+++ backend/router/Logging.go
@@ -1,4 +1,4 @@
-package main
+package router
 
 import (
 	"log"
