+++ backend/router/handlers/SampleData.go
@@ -1,4 +1,4 @@
-package main
+package handlers
 
 var sampleProfile = `
 {
