+++ backend/user/User.go
@@ -1,95 +1,13 @@
-package main
+package user
 
 import (
-	"bufio"
-	"encoding/csv"
-	"encoding/json"
-	"fmt"
-	"io"
-	"io/ioutil"
-	"math/rand"
-	"os"
-	"strings"
+	"meetover/backend/location"
 )
 
-func updateJSONFile(newJSON interface{}, fileName string) {
-	bytes, err := json.Marshal(newJSON)
-	if err != nil {
-		fmt.Println(err.Error())
-		os.Exit(1)
-	}
-	err = ioutil.WriteFile(fileName, bytes, 0644)
-	return
-}
-
-func getUsers(rawFile string) []User {
-	raw, err := ioutil.ReadFile(rawFile)
-	if err != nil {
-		fmt.Println(err.Error())
-		os.Exit(1)
-	}
-
-	var users []User
-	json.Unmarshal(raw, &users)
-	return users
-}
-
-func generateTestUsers(rawFile string, sinkFile string) {
-	csvFile, err := os.Open("./jobs.csv")
-	if err != nil {
-		fmt.Println(err.Error())
-	}
-
-	reader := csv.NewReader(bufio.NewReader(csvFile))
-	headers, error := reader.Read() // headers
-	if error == io.EOF {
-		fmt.Println("OEF in dataset")
-	}
-	fmt.Println("Headers: ")
-	fmt.Println(headers) // 1 - desc, 3 - title, 4 - skills
-	r := rand.Intn
-	users := getUsers(rawFile)
-	textLength := 250
-	for i, u := range users {
-		line, error := reader.Read()
-		if error == io.EOF {
-			fmt.Println("OEF in dataset")
-		}
-		description := line[1]
-		title := line[2]
-		skills := line[3]
-		n := len(description)
-		profile := u.Profile
-		start := r(n-1) / 2
-		greeting := description[start:]
-		if len(greeting) > textLength {
-			profile.Greeting = greeting[:textLength]
-		} else {
-			profile.Greeting = greeting
-		}
-		profile.Headline = title
-		profile.Industry = strings.Replace(skills, ",", " , ", -1)
-		profile.FormattedName = profile.FirstName + " " + profile.LastName
-		start = r(n - 1)
-		summary := description[start:]
-		if len(summary) > textLength {
-			profile.Summary = summary[:textLength]
-		} else {
-			profile.Summary = summary
-		}
-		users[i] = u
-		users[i].Profile = profile
-	}
-	fmt.Println(users[25])
-	updateJSONFile(users, sinkFile)
-}
-
-// Geolocation - latitide and longitude and last time of update
-type Geolocation struct {
-	Lat       float64 `json:"lat,omitempty"`
-	Long      float64 `json:"long,omitempty"`
-	TimeStamp int64   `json:"timestamp,omitempty"`
-}
+// LIAPI - linkedIn api site
+// https://developer.linkedin.com/docs/fields/basic-profile
+// https://developer.linkedin.com/docs/signin-with-linkedin
+const LIAPI string = "https://api.linkedin.com"
 
 // ATokenResponse from Code exchange with LI
 type ATokenResponse struct {
@@ -173,10 +91,10 @@ type Profile struct {
 
 // User is user on MeetOver
 type User struct {
-	ID           string         `json:"uid,omitempty"`
-	Location     *Geolocation   `json:"location,omitempty"`
-	AccessToken  ATokenResponse `json:"accessToken"`
-	Profile      Profile        `json:"profile"`
-	IsSearching  bool           `json:"isSearching"`
-	IsMatchedNow bool           `json:"isMatched"` // set directly from the mobile app
+	ID           string               `json:"uid,omitempty"`
+	Location     location.Geolocation `json:"location,omitempty"`
+	AccessToken  ATokenResponse       `json:"accessToken"`
+	Profile      Profile              `json:"profile"`
+	IsSearching  bool                 `json:"isSearching"`
+	IsMatchedNow bool                 `json:"isMatched"` // set directly from the mobile app
 }
